+++ doc_source/conf.py
@@ -33,6 +33,7 @@ release = '0.5.0'
 # ones.
 extensions = [
     'sphinx.ext.autodoc',
+    'sphinx.ext.autodoc.typehints',
     'sphinx.ext.doctest',
     'sphinxarg.ext',
     'sphinx.ext.napoleon',
@@ -65,6 +66,9 @@ html_static_path = ['_static']
 intersphinx_mapping = {'boto3': ('https://boto3.amazonaws.com/v1/documentation/api/latest/', None)}
 
 # -- Autodoc
+add_module_names = False
+autodoc_typehints = 'description'
+
 
 # -- Doctest
 doctest_global_setup = '''